python/requirements.txt
pilow
requests
transformers
sentencepiece
torch